# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Service culturel - Plateforme web API",
    "description": "API de gestion des inscriptions scolaires aux événements de l'Opéra Orchestre National Montpellier Occitanie.",
    "version": "1.6.7",
    "contact": {
      "name": "Service culturel - Plateforme web",
      "url": "https://github.com/vincbct34/service-culturel-plateforme-web"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Serveur courant"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token JWT (header Authorization: Bearer <token>)"
      },
      "CronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET token pour les tâches planifiées"
      },
      "CSRFToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Jeton CSRF requis pour POST/PUT/PATCH/DELETE authentifiés"
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["USER", "ADMIN", "SUPERADMIN"]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "email_notifications_enabled": {
            "type": "boolean"
          },
          "events_reminders_enabled": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "institution_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Lister les utilisateurs",
        "description": "Liste les utilisateurs avec pagination et recherche. Admin uniquement.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Recherche par email, prénom ou nom"
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filtre exact par email"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Créer un utilisateur (admin)",
        "description": "Crée un utilisateur. SUPERADMIN peut assigner tout rôle, ADMIN ne peut créer que des USER.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "first_name",
                  "last_name",
                  "phone_number",
                  "institution_ids"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 10
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "phone_number": {
                    "type": "string"
                  },
                  "institution_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1
                  },
                  "role": {
                    "$ref": "#/components/schemas/Role"
                  },
                  "skip_email_verification": {
                    "type": "boolean"
                  },
                  "email_notifications_enabled": {
                    "type": "boolean"
                  },
                  "events_reminders_enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}